view model async coroutines (#453)
Building with Kotlin and ViewModels allow us to access the
viewModelScope. We can leverage this by safe-launching computationally
expensive methods or methods that require view-model dependency.

In DDD's case, it's best to use viewModelScope for methods that interact
with our defined services.

Notes: 
- viewModelScope might be overkill for methods that just update state.
- if a top level method is called with viewModelScope, the methods
called within that are encapsulated in the same scope.