Feat/wifi aware (#480)
In this pull request, I introduced a new functionality called Wi-Fi
Aware. Currently, DDD relies on Wi-Fi Direct for communication between
BundleClient and BundleTransport. So the addition of Wi-Fi Aware
provides an extra layer of flexibility for connectivity for Android
devices that support it.

Wi-Fi Aware enables devices running Android 8.0 (API level 26) and
higher to discover and connect directly to each other without any other
type of connectivity between them. The main benefit of Wi-Fi Aware in
DDD is that devices can connect independently and without a central
coordinator, as Wi-Fi Direct requires a Group Owner. Wi-Fi Aware also
fits DDD's architecture because it is well designed for dynamic
scenarios where devices frequently joins and leaves a mobile network.

In our use case, BundleTransports (Publisher) are responsible for
publishing a service and BundleClients (Subscriber) are responsible for
discovering the published service.

My PR is the initial version of BundleClient's WiFi-Aware.