Transport Activity migration & refactor USB  (#502)
Final 'big' PR for Android migrations to Kotlin.
- Client and Transport share a lot of the same UI and ViewModel
functionalities. I wrote a general UsbViewModel class that the Client
and Transport customizations can inherit from. I also wrote a reusable
UI component that asks the user for file permissions.
- The easter egg for Transport is the 'toClient' area of the Upload
screen. Click it 7 times in <3 seconds to see the admin tabs which
consist of the log and permission screens