gradle move to AndroidApps (#595)
IntelliJ cannot handle a project with two build systems, but it does
support modules with different build systems.

this change moves all the gradle build modules into AndroidApps, so that
IntelliJ can find all the maven build and gradle built files.

this change also removes mysignal. it is abandoned at this point and
would need to be split up.