App share fix to display QR when wifi direct or internet is availableApp share fix (#710)
Fix for https://github.com/SJSU-CS-systems-group/DDD/issues/696

The QR code in App Share was disappearing when internet was turned off,
even though Wi-Fi Direct was still active. Now the QR code stays visible
as long as Wi-Fi Direct is active and an interface with IP starting with
`192.168.49.` is present.